Split files into directory, added headless mode
